Finished argparse implementation
