+++ src/studentdatabase/__init__.py
@@ -3,9 +3,7 @@ This module represents an application that is managing students.
 """
 
 from datetime import datetime
-from typing import Optional, Tuple, List
-import argparse
-from argparse import ArgumentParser, Namespace
+from typing import Tuple
 
 
 class Student:
@@ -131,129 +129,3 @@ class Student:
         """
         return (f'{self.name}, {self.birth_date}, {self.admission_yr}, {self.record_book},'
                 f' {self.group}, {self.department}, {self.faculty}, {self.gender}')
-
-
-def is_student_in_list(students: List[Student], name: str) -> int:
-    for student in students:
-        if student.name == name:
-            return students.index(student)
-    return -1
-
-
-def add_student(list_of_students: List[Student]):
-    student = Student()
-    student.set_student()
-    list_of_students.append(student)
-
-
-def get_existing_student(list_of_students: List[Student]):
-    student_in_the_list = False
-    while not student_in_the_list:
-        answer = input("Please enter student´s name: ")
-        student_index = is_student_in_list(list_of_students, answer)
-        if student_index > -1:
-            print(list_of_students[student_index].get_student())
-            student_in_the_list = True
-        else:
-            print("Student not found!")
-
-
-def delete_student(list_of_students: List[Student]):
-    student_in_the_list = False
-    while not student_in_the_list:
-        answer = input("Please enter student´s name: ")
-        student_index = is_student_in_list(list_of_students, answer)
-        if student_index > -1:
-            list_of_students.pop(student_index)
-            student_in_the_list = True
-        else:
-            print("Student not found!")
-
-
-def main():
-    list_of_students: List[Student] = []
-    users_answer = -1
-    while users_answer != 5:
-        print("""Hello!
-This is student´s database!
-What shall we do?
-
-1. Add a student
-2. Get an existing student
-3. Delete a student
-4. Show all students
-5. End the program
-""")
-
-        users_correct_answer = False
-        while not users_correct_answer:
-            user = input("Please enter your choice [1-5]: ")
-            if user.isdigit() and 0 < int(user) < 6:
-                users_correct_answer = True
-                users_answer = int(user)
-            else:
-                print("Not a number between 1 and 5!")
-
-        if users_answer == 1:
-            add_student(list_of_students)
-            print("""
-Student succesfully added! 
-            
-/stop to stop the code""")
-
-            user_menu = str(input())
-            if user_menu == "/stop":
-                print("code stopped.")
-                break
-
-        elif users_answer == 2:
-            get_existing_student(list_of_students)
-
-        elif users_answer == 3:
-            delete_student(list_of_students)
-
-        elif users_answer == 4:
-            print(f'''The list of the students:
-{list_of_students}''')
-
-        elif users_answer == 5:
-            print("Goodbye!")
-
-
-if __name__ == "__main__":
-    main()
-    parser = argparse.ArgumentParser(prog="interactive user input",
-                                     description="This module interacts users input.",
-                                     epilog="End of the arguments.")
-
-    parser.add_argument("operation", choices=["add", "interactive"],
-                        help="Choose an operation: <add> to add a student or <interactive> to see more options.")
-
-    # parser.add_argument("--name", "--birthday", "--admission_year", "--record_book",
-    #                     "--group", "--department", "--faculty", "--gender", help="Adds a student into database.")
-
-    parser.add_argument("-nm=", "--name=", help="Adds a student into database.")
-    parser.add_argument("-bd=", "--birthday=", help="Adds a student into database.")
-    parser.add_argument("-a_y=", "--admission_year=", help="Adds a student into database.")
-    parser.add_argument("-r_b=", "--record_book=", help="Adds a student into database.")
-    parser.add_argument("-gr=", "--group=", help="Adds a student into database.")
-    parser.add_argument("-dp=", "--department=", help="Adds a student into database.")
-    parser.add_argument("-fa=", "--faculty=", help="Adds a student into database.")
-    parser.add_argument("-gd=","--gender=", help="Adds a student into database.")
-
-    args = parser.parse_args()
-    list_of_students: List[Student] = []
-
-    if args.operation == "add":
-
-        add_student(list_of_students)
-        print("Student succesfully added!")
-
-
-
-
-
-
-
-
-
